spindle-project/code/code/requirements.txt
# Core scientific stack
numpy
scipy
pandas
matplotlib
seaborn

# Deep learning
torch>=2.0
torchvision
torchaudio

# EEG and neuroimaging
mne

# Config and experiment tracking
pyyaml
omegaconf
wandb

# Utilities
tqdm
scikit-learn
joblib

# For downloading and handling data
requests

